+++ androidx-kt/src/main/java/me/panpf/androidxkt/os/storage/Storagex.kt
@@ -38,14 +38,29 @@ import java.io.File
 inline fun File.getFreeBytes(): Long = Storagex.getFreeBytes(this)
 
 /**
+ * Get the number of free bytes of the given path. The directory does not exist and the creation is unsuccessful. [defaultValue]
+ */
+inline fun File.getFreeBytes(defaultValue: Long): Long = Storagex.getFreeBytes(this, defaultValue)
+
+/**
  * Get the number of total bytes of the given path
  */
 inline fun File.getTotalBytes(): Long = Storagex.getTotalBytes(this)
 
 /**
+ * Get the number of total bytes of the given path. The directory does not exist and the creation is unsuccessful. [defaultValue]
+ */
+inline fun File.getTotalBytes(defaultValue: Long): Long = Storagex.getTotalBytes(this, defaultValue)
+
+/**
  * Get the number of available bytes of the given path
  */
 inline fun File.getAvailableBytes(): Long = Storagex.getAvailableBytes(this)
+
+/**
+ * Get the number of available bytes of the given path. The directory does not exist and the creation is unsuccessful. [defaultValue]
+ */
+inline fun File.getAvailableBytes(defaultValue: Long): Long = Storagex.getAvailableBytes(this, defaultValue)
 
 /**
  * Get the number of free bytes for the primary shared/external storage media